assignment3/requirements.txt
requests
pandas
sqlalchemy
streamlit
sqlite3
python-dotenv